Extract test of Data package.
